:recycle: Dynamically generate tests for nacl libraries
