+++ packages/lisk-cryptography/test/nacl/index.js
@@ -1,8 +1,21 @@
+/*
+ * Copyright © 2018 Lisk Foundation
+ *
+ * See the LICENSE file at the top-level directory of this distribution
+ * for licensing information.
+ *
+ * Unless otherwise agreed in a custom licensing agreement with the Lisk Foundation,
+ * no part of this software, including this file, may be copied, modified,
+ * propagated, or distributed except according to the terms contained in the
+ * LICENSE file.
+ *
+ * Removal or modification of this copyright notice is prohibited.
+ *
+ */
 import * as fast from '../../src/nacl/fast';
 import * as slow from '../../src/nacl/slow';
 // Require is used for stubbing
-// eslint-disable-next-line
-const _module = require('module');
+const moduleLibrary = require('module');
 
 const resetTest = () => {
 	// Reset environment variable
@@ -11,18 +24,15 @@ const resetTest = () => {
 	delete require.cache[require.resolve('../../src/nacl/index')];
 };
 
-const removeConstants = lib => {
+const stripConstants = library => {
 	// Constants are added in /src/index.js
-	// eslint-disable-next-line
-	delete lib.NACL_SIGN_PUBLICKEY_LENGTH;
-	// eslint-disable-next-line
-	delete lib.NACL_SIGN_SIGNATURE_LENGTH;
-	return lib;
+	const strippedLib = Object.assign({}, library);
+	delete strippedLib.NACL_SIGN_PUBLICKEY_LENGTH;
+	delete strippedLib.NACL_SIGN_SIGNATURE_LENGTH;
+	return strippedLib;
 };
 
 describe('nacl index.js', () => {
-	let loadedLibrary;
-	// Store current env variable and set it back after tests are run
 	let initialEnvVar;
 	before(() => {
 		initialEnvVar = process.env.NACL_FAST;
@@ -43,14 +53,6 @@ describe('nacl index.js', () => {
 		return Promise.resolve();
 	});
 
-	it('should load nacl slow if process.env.NACL_FAST is set to disable', () => {
-		process.env.NACL_FAST = 'disable';
-		// eslint-disable-next-line
-		loadedLibrary = require('../../src/nacl/index');
-		loadedLibrary = removeConstants(loadedLibrary);
-		return expect(loadedLibrary).to.be.eql(slow);
-	});
-
 	describe('nacl fast installed', () => {
 		beforeEach(() => {
 			resetTest();
@@ -59,26 +61,26 @@ describe('nacl index.js', () => {
 
 		it('should load nacl fast if process.env.NACL_FAST is set to enable', () => {
 			process.env.NACL_FAST = 'enable';
-			// eslint-disable-next-line
-			loadedLibrary = require('../../src/nacl/index');
-			loadedLibrary = removeConstants(loadedLibrary);
-			return expect(loadedLibrary).to.be.eql(fast);
+			// eslint-disable-next-line global-require
+			const loadedLibrary = require('../../src/nacl');
+			const strippedLibrary = stripConstants(loadedLibrary);
+			return expect(strippedLibrary).to.be.eql(fast);
 		});
 
 		it('should load nacl slow if process.env.NACL_FAST is set to disable', () => {
 			process.env.NACL_FAST = 'disable';
-			// eslint-disable-next-line
-			loadedLibrary = require('../../src/nacl/index');
-			loadedLibrary = removeConstants(loadedLibrary);
-			return expect(loadedLibrary).to.be.eql(slow);
+			// eslint-disable-next-line global-require
+			const loadedLibrary = require('../../src/nacl');
+			const strippedLibrary = stripConstants(loadedLibrary);
+			return expect(strippedLibrary).to.be.eql(slow);
 		});
 
 		it('should load nacl fast if process.env.NACL_FAST is undefined', () => {
 			process.env.NACL_FAST = undefined;
-			// eslint-disable-next-line
-			loadedLibrary = require('../../src/nacl/index');
-			loadedLibrary = removeConstants(loadedLibrary);
-			return expect(loadedLibrary).to.be.eql(fast);
+			// eslint-disable-next-line global-require
+			const loadedLibrary = require('../../src/nacl');
+			const strippedLibrary = stripConstants(loadedLibrary);
+			return expect(strippedLibrary).to.be.eql(fast);
 		});
 	});
 
@@ -88,35 +90,41 @@ describe('nacl index.js', () => {
 			resetTest();
 			// "require" is a wrapper around Module._load which handles the actual loading
 			sandbox
-				.stub(_module, '_load')
+				.stub(moduleLibrary, '_load')
 				.callThrough()
 				.withArgs('./fast')
 				.throws(moduleNotFoundError);
 			return Promise.resolve();
 		});
 
+		it('should set process.env.NACL_FAST to disable', () => {
+			// eslint-disable-next-line global-require
+			require('../../src/nacl');
+			return expect(process.env.NACL_FAST).to.eql('disable');
+		});
+
 		it('should load nacl slow if process.env.NACL_FAST is set to enable', () => {
 			process.env.NACL_FAST = 'enable';
-			// eslint-disable-next-line
-			loadedLibrary = require('../../src/nacl/index');
-			loadedLibrary = removeConstants(loadedLibrary);
-			return expect(loadedLibrary).to.eql(slow);
+			// eslint-disable-next-line global-require
+			const loadedLibrary = require('../../src/nacl');
+			const strippedLibrary = stripConstants(loadedLibrary);
+			return expect(strippedLibrary).to.eql(slow);
 		});
 
 		it('should load nacl slow if process.env.NACL_FAST is set to disable', () => {
 			process.env.NACL_FAST = 'disable';
-			// eslint-disable-next-line
-			loadedLibrary = require('../../src/nacl/index');
-			loadedLibrary = removeConstants(loadedLibrary);
-			return expect(loadedLibrary).to.eql(slow);
+			// eslint-disable-next-line global-require
+			const loadedLibrary = require('../../src/nacl');
+			const strippedLibrary = stripConstants(loadedLibrary);
+			return expect(strippedLibrary).to.eql(slow);
 		});
 
 		it('should load nacl slow if process.env.NACL_FAST is undefined', () => {
 			process.env.NACL_FAST = undefined;
-			// eslint-disable-next-line
-			loadedLibrary = require('../../src/nacl/index');
-			loadedLibrary = removeConstants(loadedLibrary);
-			return expect(loadedLibrary).to.eql(slow);
+			// eslint-disable-next-line global-require
+			const loadedLibrary = require('../../src/nacl');
+			const strippedLibrary = stripConstants(loadedLibrary);
+			return expect(strippedLibrary).to.eql(slow);
 		});
 	});
 });
